# openapi-schema.yaml
openapi: 3.0.3
info:
  title: Bus Kiosk Backend API
  version: 1.0.0
  description: Industrial REST API for Bus Kiosk face recognition system
paths:
...
  /api/v1/kiosks/{kiosk_id}/check-updates/:
    get:
      operationId: kiosk_check_updates
      description: Check if kiosk needs database update
      parameters:
      - in: path
        name: kiosk_id
        schema:
          type: string
        required: true
      - in: query
        name: last_sync_hash
        schema:
          type: string
        description: Last content hash from kiosk
      tags:
      - api
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckUpdatesResponse'
          description: ''
  /api/v1/kiosks/{kiosk_id}/heartbeat/:
    post:
      operationId: kiosk_heartbeat
      description: Report kiosk health and sync status
      parameters:
      - in: path
        name: kiosk_id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Heartbeat'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Heartbeat'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Heartbeat'
        required: true
      security:
      - Bearer: []
      responses:
        '204':
          description: No response body
  /api/v1/kiosks/{kiosk_id}/location/:
    post:
      operationId: kiosk_update_location
      description: Update bus GPS location. Kiosk sends location when bus moves significantly
        or every 2 minutes.
      parameters:
      - in: path
        name: kiosk_id
        schema:
          type: string
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusLocation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BusLocation'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BusLocation'
        required: true
      security:
      - Bearer: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusLocation'
          description: ''
  /api/v1/kiosks/{kiosk_id}/snapshot/:
    get:
      operationId: kiosk_download_snapshot
      description: Download kiosk database snapshot (binary SQLite file). Returns
        raw binary data with x-snapshot-checksum header for verification.
      parameters:
      - in: path
        name: kiosk_id
        schema:
          type: string
        required: true
      tags:
      - api
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          description: ''
...
  /api/v1/kiosks/logs/:
    post:
      operationId: kiosk_log
      description: Kiosk logging endpoint for device log submission
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceLog'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeviceLog'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeviceLog'
        required: true
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
          description: ''
...
  /api/v1/logs/:
    get:
      operationId: api_v1_logs_list
      description: Read-only ViewSet for device logs (any authenticated user)
      parameters:
      - in: query
        name: kiosk
        schema:
          type: string
        description: Unique kiosk device identifier (e.g., KIOSK001, BUS123-KIOSK)
      - in: query
        name: log_level
        schema:
          type: string
          x-spec-enum-id: 7b1a6f8785c20640
          enum:
          - CRITICAL
          - DEBUG
          - ERROR
          - INFO
          - WARN
        description: |-
          Log level severity

          * `DEBUG` - Debug
          * `INFO` - Info
          * `WARN` - Warning
          * `ERROR` - Error
          * `CRITICAL` - Critical
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - in: query
        name: timestamp
        schema:
          type: string
          format: date-time
      tags:
      - api
      security:
      - cookieAuth: []
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDeviceLogList'
          description: ''
...
  /api/v1/logs/summary/:
    get:
      operationId: api_v1_logs_summary_retrieve
      description: Get logs summary by level and time
      tags:
      - api
      security:
      - cookieAuth: []
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceLog'
          description: ''
...
components:
  schemas:
...
    BusLocation:
      type: object
      description: Serializer for GPS location updates from kiosk
      properties:
        location_id:
          type: integer
          readOnly: true
          description: Auto-incrementing location entry ID
        latitude:
          type: number
          format: double
          description: GPS latitude coordinate
        longitude:
          type: number
          format: double
          description: GPS longitude coordinate
        accuracy:
          type: number
          format: double
          nullable: true
          description: GPS accuracy in meters
        speed:
          type: number
          format: double
          nullable: true
          description: Speed in km/h
        heading:
          type: number
          format: double
          nullable: true
          description: Heading/bearing in degrees (0-360)
        timestamp:
          type: string
          format: date-time
          description: When this location was recorded by the kiosk
      required:
      - latitude
      - location_id
      - longitude
      - timestamp
    CheckUpdatesResponse:
      type: object
      description: Serializer for check updates response
      properties:
        needs_update:
          type: boolean
          description: Whether kiosk needs to download new snapshot
        current_version:
          type: string
          description: Current database version timestamp
        student_count:
          type: integer
          description: Number of students for this bus
        embedding_count:
          type: integer
          description: Number of embeddings for this bus
        content_hash:
          type: string
          description: Content hash for integrity verification
      required:
      - content_hash
      - current_version
      - embedding_count
      - needs_update
      - student_count
...
    DeviceLog:
      type: object
      description: Serializer for device logs
      properties:
        log_id:
          type: integer
          readOnly: true
          description: Auto-incrementing log entry ID
        kiosk:
          type: string
          description: Kiosk that generated this log entry
        kiosk_id:
          type: string
          readOnly: true
        log_level:
          enum:
          - DEBUG
          - INFO
          - WARN
          - ERROR
          - CRITICAL
          type: string
          x-spec-enum-id: 7b1a6f8785c20640
          description: |-
            Log level severity

            * `DEBUG` - Debug
            * `INFO` - Info
            * `WARN` - Warning
            * `ERROR` - Error
            * `CRITICAL` - Critical
        message:
          type: string
          description: Log message content
        metadata:
          description: Additional structured data as JSON
        timestamp:
          type: string
          format: date-time
          readOnly: true
          description: When this log entry was created
      required:
      - kiosk
      - kiosk_id
      - log_id
      - log_level
      - message
      - timestamp
...
    HealthData:
      type: object
      description: Serializer for health metrics
      properties:
        battery_level:
          type: integer
          maximum: 100
          minimum: 0
          nullable: true
        is_charging:
          type: boolean
          default: false
        storage_available_mb:
          type: integer
          minimum: 0
          nullable: true
        camera_active:
          type: boolean
          default: false
        network_type:
          type: string
          nullable: true
          maxLength: 20
        app_version:
          type: string
          nullable: true
          maxLength: 20
        last_face_detected_ago_min:
          type: integer
          minimum: 0
          nullable: true
        faces_detected_today:
          type: integer
          minimum: 0
          default: 0
        students_identified_today:
          type: integer
          minimum: 0
          default: 0
    Heartbeat:
      type: object
      description: Serializer for kiosk heartbeat
      properties:
        timestamp:
          type: string
          format: date-time
          description: Heartbeat timestamp
        database_version:
          type: string
          description: Current database version on kiosk
          maxLength: 50
        database_hash:
          type: string
          description: Database content hash
          maxLength: 64
        student_count:
          type: integer
          minimum: 0
          description: Students in DB
        embedding_count:
          type: integer
          minimum: 0
          description: Embeddings in DB
        health:
          $ref: '#/components/schemas/HealthData'
      required:
      - database_version
      - embedding_count
      - student_count
      - timestamp
...
    PaginatedDeviceLogList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeviceLog'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.easypool.in
  description: Production API